Fix avatar source URL for subpages in pelican-bootstrap3
When going to a subpage, such as a category page, my avatar image was not
being displayed when using the suggested format 'images/profile.png' from
the readme. The URL turned into 'category/images/profile.png' which doesn't
exist.

Include the site URL so that the image appears on all pages. This looked
like the convention used in other places like favicon.